requirements.txt
langchain
pyyaml
chromadb
openai
flask
Flask-Session
sentence-transformers